# Buffer API changes, Optimization and refactor of ResizeProcessors