--- a/src/ImageSharp.Drawing/Brushes/ImageBrush{TColor}.cs
+++ b/src/ImageSharp.Drawing/Brushes/ImageBrush{TColor}.cs
@@ -119,7 +119,7 @@ namespace ImageSharp.Drawing.Brushes
 
                 using (PinnedBuffer<float> buffer = new PinnedBuffer<float>(scanlineBuffer))
                 {
-                    BufferPointer<float> slice = buffer.Slice(offset);
+                    BufferSpan<float> slice = buffer.Slice(offset);
 
                     for (int xPos = 0; xPos < scanlineWidth; xPos++)
                     {