--- a/src/ImageSharp.Drawing/Brushes/PatternBrush{TColor}.cs
+++ b/src/ImageSharp.Drawing/Brushes/PatternBrush{TColor}.cs
@@ -152,7 +152,7 @@ namespace ImageSharp.Drawing.Brushes
 
                 using (PinnedBuffer<float> buffer = new PinnedBuffer<float>(scanlineBuffer))
                 {
-                    BufferPointer<float> slice = buffer.Slice(offset);
+                    BufferSpan<float> slice = buffer.Slice(offset);
 
                     for (int xPos = 0; xPos < scanlineWidth; xPos++)
                     {