--- a/src/ImageSharp.Drawing/Brushes/Processors/BrushApplicator.cs
+++ b/src/ImageSharp.Drawing/Brushes/Processors/BrushApplicator.cs
@@ -57,7 +57,7 @@ namespace ImageSharp.Drawing.Processors
 
             using (PinnedBuffer<float> buffer = new PinnedBuffer<float>(scanlineBuffer))
             {
-                BufferPointer<float> slice = buffer.Slice(offset);
+                BufferSpan<float> slice = buffer.Slice(offset);
 
                 for (int xPos = 0; xPos < scanlineWidth; xPos++)
                 {