--- a/src/ImageSharp/Common/Memory/IPinnedImageBuffer{T}.cs
+++ b/src/ImageSharp/Common/Memory/IPinnedImageBuffer{T}.cs
@@ -1,0 +1,31 @@
+// <copyright file="IPinnedImageBuffer{T}.cs" company="James Jackson-South">
+// Copyright (c) James Jackson-South and contributors.
+// Licensed under the Apache License, Version 2.0.
+// </copyright>
+
+namespace ImageSharp
+{
+    /// <summary>
+    /// An interface that represents a pinned buffer of value type objects
+    /// interpreted as a 2D region of <see cref="Width"/> x <see cref="Height"/> elements.
+    /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    internal interface IPinnedImageBuffer<T>
+        where T : struct
+    {
+        /// <summary>
+        /// Gets the width.
+        /// </summary>
+        int Width { get; }
+
+        /// <summary>
+        /// Gets the height.
+        /// </summary>
+        int Height { get; }
+
+        /// <summary>
+        /// Gets a <see cref="BufferSpan{T}"/> to the backing buffer.
+        /// </summary>
+        BufferSpan<T> Span { get; }
+    }
+}