--- a/src/ImageSharp/Processing/Processors/Transforms/ResamplingWeightedProcessor.cs
+++ b/src/ImageSharp/Processing/Processors/Transforms/ResamplingWeightedProcessor.cs
@@ -156,6 +156,5 @@ namespace ImageSharp.Processing.Processors
 
             return result;
         }
-
     }
 }