… src/ImageSharp/Common/Memory/BufferPointer{T}.cs
… src/ImageSharp/Common/Memory/PinnedBuffer{T}.cs
… src/ImageSharp/Common/Memory/PinnedImageBufferExtensions.cs
… src/ImageSharp/Image/PixelAccessor{TColor}.cs
… src/ImageSharp/Image/PixelArea{TColor}.cs
… src/ImageSharp/Processing/Processors/Transforms/CompandingResizeProcessor.cs
… src/ImageSharp/Processing/Processors/Transforms/ResamplingWeightedProcessor.Weights.cs
… src/ImageSharp/Processing/Processors/Transforms/ResizeProcessor.cs
… src/ImageSharp/Processing/Transforms/Resize.cs
… tests/ImageSharp.Sandbox46/Program.cs
… tests/ImageSharp.Tests/Colors/BulkPixelOperationsTests.cs
… tests/ImageSharp.Tests/Common/BufferSpanTests.cs
… tests/ImageSharp.Tests/Common/PinnedBufferTests.cs
… tests/ImageSharp.Tests/Common/PinnedImageBufferTests.cs
… tests/ImageSharp.Tests/Common/TestStructs.cs
… tests/ImageSharp.Tests/Processors/Filters/ResizeProfilingBenchmarks.cs
… tests/ImageSharp.Tests/Processors/Filters/ResizeTests.cs