@antonfirsov 2017-03-20
I think it's enough to move out if-s from the inner loop. We can benchmark a version where they are outside the outer loops, but I don't think it's worth the efforts.